+++ lib/src/controls/form/text_box.dart
@@ -1007,9 +1007,7 @@ class _TextBoxState extends State<TextBox>
         themeData.inactiveColor;
 
     final selectionColor = DefaultSelectionStyle.of(context).selectionColor ??
-        themeData.accentColor
-            .defaultBrushFor(themeData.brightness)
-            .withOpacity(0.2);
+        themeData.accentColor.normal;
 
     // Set configuration as disabled if not otherwise specified. If specified,
     // ensure that configuration uses Fluent text style for misspelled words